Fix help section rendering of Help editor module Markdown content
