+++ htdocs/js/components/helpHandler.js
@@ -1,0 +1,2 @@
+!function(modules){function __webpack_require__(moduleId){if(installedModules[moduleId])return installedModules[moduleId].exports;var module=installedModules[moduleId]={exports:{},id:moduleId,loaded:!1};return modules[moduleId].call(module.exports,module,module.exports,__webpack_require__),module.loaded=!0,module.exports}var installedModules={};return __webpack_require__.m=modules,__webpack_require__.c=installedModules,__webpack_require__.p="",__webpack_require__(0)}([function(module,exports){"use strict";function _classCallCheck(instance,Constructor){if(!(instance instanceof Constructor))throw new TypeError("Cannot call a class as a function")}function _possibleConstructorReturn(self,call){if(!self)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!call||"object"!=typeof call&&"function"!=typeof call?self:call}function _inherits(subClass,superClass){if("function"!=typeof superClass&&null!==superClass)throw new TypeError("Super expression must either be null or a function, not "+typeof superClass);subClass.prototype=Object.create(superClass&&superClass.prototype,{constructor:{value:subClass,enumerable:!1,writable:!0,configurable:!0}}),superClass&&(Object.setPrototypeOf?Object.setPrototypeOf(subClass,superClass):subClass.__proto__=superClass)}var _createClass=function(){function defineProperties(target,props){for(var i=0;i<props.length;i++){var descriptor=props[i];descriptor.enumerable=descriptor.enumerable||!1,descriptor.configurable=!0,"value"in descriptor&&(descriptor.writable=!0),Object.defineProperty(target,descriptor.key,descriptor)}}return function(Constructor,protoProps,staticProps){return protoProps&&defineProperties(Constructor.prototype,protoProps),staticProps&&defineProperties(Constructor,staticProps),Constructor}}(),DisplayContent=function(_React$Component){function DisplayContent(props){return _classCallCheck(this,DisplayContent),_possibleConstructorReturn(this,(DisplayContent.__proto__||Object.getPrototypeOf(DisplayContent)).call(this,props))}return _inherits(DisplayContent,_React$Component),_createClass(DisplayContent,[{key:"componentDidMount",value:function(){console.log(React.createElement(DisplayContent,{Topic:content.topic,Content:content.content,Updated:content.updated}))}},{key:"render",value:function(){function HasDate(date){return React.createElement("div",null,React.createElement("hr",null),React.createElement("p",null,"Last updated: ",date))}return React.createElement("div",null,React.createElement("hr",{id:"help-separator"}),React.createElement("h3",null,this.props.Topic?this.props.Topic:""),React.createElement(Markdown,{content:this.props.Content}),this.props.Updated&&HasDate(this.props.Updated))}}]),DisplayContent}(React.Component);$(document).ready(function(){$("#menu-toggle").click(function(e){e.preventDefault(),$(".wrapper").toggleClass("active")}),$(".help-button").click(function(e){var helpContent=$("div.help-content");if(helpContent.length)return helpContent.toggle(),void e.preventDefault();var getParams={};getParams.testName=loris.TestName,""!==loris.Subtest&&(getParams.subtest=loris.Subtest),document.cookie="LastUrl=; expires=Thu, 01-Jan-70 00:00:01 GMT;",$.get(loris.BaseURL+"/help_editor/ajax/help.php",getParams,function(content){var wrap,div=document.createElement("div"),btn=document.createElement("BUTTON"),edit=document.createElement("BUTTON"),text=document.createTextNode("Edit"),button=document.createTextNode("Close");new String;"markdown"===content.format?(wrap=document.createElement("div"),wrap.setAttribute("id","help-wrapper"),ReactDOM.render(RMarkdown({content:content.content}),wrap)):(wrap=document.createElement("pre"),wrap.setAttribute("id","help-wrapper"),ReactDOM.render(React.createElement(DisplayContent,{Topic:content.topic,Content:content.content,Updated:content.updated}),wrap)),btn.appendChild(button),btn.className="btn btn-default",btn.setAttribute("id","helpclose"),edit.appendChild(text),edit.className="btn btn-default",edit.setAttribute("id","helpedit"),div.appendChild(wrap),div.appendChild(btn),loris.userHasPermission("context_help")&&"markdown"!==content.format&&(div.appendChild(edit),edit.addEventListener("click",function(e){e.preventDefault(),document.cookie="LastUrl = "+document.location.toString(),window.open(loris.BaseURL+"/help_editor/edit_help_content/?section="+getParams.testName+"&subsection="+getParams.subtest,"_self")})),document.getElementById("page").appendChild(div),div.setAttribute("class","help-content"),$(div).addClass("visible"),btn.addEventListener("click",function(e){$(div).hide(),e.preventDefault()}),e.preventDefault()},"json")}),$(".dynamictable").DynamicTable(),$(".fileUpload").FileUpload(),swal.setDefaults({confirmButtonColor:"#064785"})}),$(document).ajaxError(function(event,jqxhr,settings,thrownError){401===jqxhr.status&&($("#login-modal").hasClass("in")?$("#login-modal-error").show():($("#login-modal").modal("show"),$("#modal-login").click(function(e){e.preventDefault();var data={username:$("#modal-username").val(),password:$("#modal-password").val(),login:"Login"};$.ajax({type:"post",url:loris.BaseURL+"/main.php",data:data,success:function(){$("#login-modal-error").hide(),$("#login-modal").modal("hide")}})})))})}]);
+//# sourceMappingURL=helpHandler.js.map
