+++ modules/help_editor/js/help_editor_helper.js
@@ -1,2 +1,2 @@
-!function(modules){function __webpack_require__(moduleId){if(installedModules[moduleId])return installedModules[moduleId].exports;var module=installedModules[moduleId]={exports:{},id:moduleId,loaded:!1};return modules[moduleId].call(module.exports,module,module.exports,__webpack_require__),module.loaded=!0,module.exports}var installedModules={};return __webpack_require__.m=modules,__webpack_require__.c=installedModules,__webpack_require__.p="",__webpack_require__(0)}([function(module,exports){"use strict";function _classCallCheck(instance,Constructor){if(!(instance instanceof Constructor))throw new TypeError("Cannot call a class as a function")}function _possibleConstructorReturn(self,call){if(!self)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!call||"object"!=typeof call&&"function"!=typeof call?self:call}function _inherits(subClass,superClass){if("function"!=typeof superClass&&null!==superClass)throw new TypeError("Super expression must either be null or a function, not "+typeof superClass);subClass.prototype=Object.create(superClass&&superClass.prototype,{constructor:{value:subClass,enumerable:!1,writable:!0,configurable:!0}}),superClass&&(Object.setPrototypeOf?Object.setPrototypeOf(subClass,superClass):subClass.__proto__=superClass)}var _createClass=function(){function defineProperties(target,props){for(var i=0;i<props.length;i++){var descriptor=props[i];descriptor.enumerable=descriptor.enumerable||!1,descriptor.configurable=!0,"value"in descriptor&&(descriptor.writable=!0),Object.defineProperty(target,descriptor.key,descriptor)}}return function(Constructor,protoProps,staticProps){return protoProps&&defineProperties(Constructor.prototype,protoProps),staticProps&&defineProperties(Constructor,staticProps),Constructor}}(),EditContent=function(_React$Component){function EditContent(props){return _classCallCheck(this,EditContent),_possibleConstructorReturn(this,(EditContent.__proto__||Object.getPrototypeOf(EditContent)).call(this,props))}return _inherits(EditContent,_React$Component),_createClass(EditContent,[{key:"render",value:function(){function removeDiv(event){document.getElementById("preview").remove(),event.preventDefault()}return React.createElement("div",null,React.createElement("pre",null,React.createElement("h3",null,this.props.Title),React.createElement("p",null,React.createElement(Markdown,{content:this.props.Content})),React.createElement("hr",null),React.createElement("p",null,"Last updated: ",this.props.MyDate.getFullYear(),"-",this.props.MyDate.getMonth()+1,"-",this.props.MyDate.getDate()," ",this.props.MyDate.getHours(),":",this.props.MyDate.getMinutes(),":",this.props.MyDate.getSeconds())),React.createElement("button",{className:"btn btn-default",id:"helpclose",onClick:removeDiv},"Close"))}}]),EditContent}(React.Component);$(document).ready(function(){$("input[name=preview]").click(function(e){$("div.help-content").length&&($("div.help-content").remove(),e.preventDefault());var title=$('input[name="title"]').val(),content=$('textarea[name="content"]').val(),myDate=new Date,div=document.createElement("div");document.createTextNode("Edit");document.getElementById("page").appendChild(div),div.setAttribute("class","help-content"),div.setAttribute("id","preview"),ReactDOM.render(React.createElement(EditContent,{Title:title,Content:content,MyDate:myDate}),document.getElementById("preview")),e.preventDefault()})})}]);
+!function(modules){function __webpack_require__(moduleId){if(installedModules[moduleId])return installedModules[moduleId].exports;var module=installedModules[moduleId]={exports:{},id:moduleId,loaded:!1};return modules[moduleId].call(module.exports,module,module.exports,__webpack_require__),module.loaded=!0,module.exports}var installedModules={};return __webpack_require__.m=modules,__webpack_require__.c=installedModules,__webpack_require__.p="",__webpack_require__(0)}([function(module,exports){"use strict";function _classCallCheck(instance,Constructor){if(!(instance instanceof Constructor))throw new TypeError("Cannot call a class as a function")}function _possibleConstructorReturn(self,call){if(!self)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!call||"object"!=typeof call&&"function"!=typeof call?self:call}function _inherits(subClass,superClass){if("function"!=typeof superClass&&null!==superClass)throw new TypeError("Super expression must either be null or a function, not "+typeof superClass);subClass.prototype=Object.create(superClass&&superClass.prototype,{constructor:{value:subClass,enumerable:!1,writable:!0,configurable:!0}}),superClass&&(Object.setPrototypeOf?Object.setPrototypeOf(subClass,superClass):subClass.__proto__=superClass)}var _createClass=function(){function defineProperties(target,props){for(var i=0;i<props.length;i++){var descriptor=props[i];descriptor.enumerable=descriptor.enumerable||!1,descriptor.configurable=!0,"value"in descriptor&&(descriptor.writable=!0),Object.defineProperty(target,descriptor.key,descriptor)}}return function(Constructor,protoProps,staticProps){return protoProps&&defineProperties(Constructor.prototype,protoProps),staticProps&&defineProperties(Constructor,staticProps),Constructor}}(),EditContent=function(_React$Component){function EditContent(props){return _classCallCheck(this,EditContent),_possibleConstructorReturn(this,(EditContent.__proto__||Object.getPrototypeOf(EditContent)).call(this,props))}return _inherits(EditContent,_React$Component),_createClass(EditContent,[{key:"componentDidMount",value:function(){console.log(JSON.stringify(this.props.Content))}},{key:"render",value:function(){function removeDiv(event){document.getElementById("preview").remove(),event.preventDefault()}return React.createElement("div",null,React.createElement("pre",null,React.createElement("h3",null,this.props.Title),React.createElement(Markdown,{content:this.props.Content}),React.createElement("hr",null),React.createElement("p",null,"Last updated: ",this.props.MyDate.getFullYear(),"-",this.props.MyDate.getMonth()+1,"-",this.props.MyDate.getDate()," ",this.props.MyDate.getHours(),":",this.props.MyDate.getMinutes(),":",this.props.MyDate.getSeconds())),React.createElement("button",{className:"btn btn-default",id:"helpclose",onClick:removeDiv},"Close"))}}]),EditContent}(React.Component);$(document).ready(function(){$("input[name=preview]").click(function(e){$("div.help-content").length&&($("div.help-content").remove(),e.preventDefault());var title=$('input[name="title"]').val(),content=$('textarea[name="content"]').val(),myDate=new Date,div=document.createElement("div");document.createTextNode("Edit");document.getElementById("page").appendChild(div),div.setAttribute("class","help-content"),div.setAttribute("id","preview"),ReactDOM.render(React.createElement(EditContent,{Title:title,Content:content,MyDate:myDate}),document.getElementById("preview")),e.preventDefault()})})}]);
 //# sourceMappingURL=help_editor_helper.js.map
