+++ php/libraries/NDB_Page.class.inc
@@ -674,7 +674,6 @@ class NDB_Page
 
         $files = array(
                   $baseurl . '/js/jquery/jquery-1.11.0.min.js',
-                  $baseurl . '/js/helpHandler.js',
                   $baseurl . '/js/modernizr/modernizr.min.js',
                   // Only load minified file on production, not sandboxes
                   $baseurl . '/js/polyfills.js',
@@ -689,6 +688,7 @@ class NDB_Page
                   $baseurl . "/js/util/queryString.js",
                   $baseurl . '/js/components/Form.js',
                   $baseurl . '/js/components/Markdown.js',
+                  $baseurl . '/js/helpHandler.js',
                  );
         return $files;
     }
